Discovery Bid Adapter : fix window.top bug
